refactored TrafficIncident
created dto package and refactored TrafficIncident into dto package
